feat(#220): Error / Loading 상태일때 UI 적용
